application: Keep startup-id from platform data around
Since glib commit 1bb48a76a6 `get_platform_data` is also called
for non-remote instances. Get the `startup_notification_id` from the
passed in data as otherwise we might send an old activation token and
activation will fail.

This can be reproduced by starting e.g. gnome-control-center and then
activating an action via DBus like:

```
$ WAYLAND_DEBUG=1 DESKTOP_STARTUP_ID=asdfsadfadaf gnome-control-center |& grep "xdg_activation.*activate"
$ gdbus call --session --dest org.gnome.Settings \
                       --object-path /org/gnome/Settings \
                       --method org.freedesktop.Application.ActivateAction \
		       'launch-panel' \
		       "[<('bluetooth', @av [])>]" \
		       "{'activation-token': <'c74447bffb17e98a846b73a3a815d285'>, 'desktop-startup-id': <'c74447bffb17e98a846b73a3a815d285'>}"
```

Before this change we send the outdated token when activating the
action:

```
[4183012.886] {Default Queue}  -> xdg_activation_v1#17.activate("asdfsadfadaf", wl_surface#43)
[4188324.543] {Default Queue}  -> xdg_activation_v1#17.activate("asdfsadfadaf", wl_surface#43)
```

After we send the one supplied via `platform-data`:

```
[3636045.303] {Default Queue}  -> xdg_activation_v1#17.activate("asdfsadfadaf", wl_surface#43)
[4136885.891] {Default Queue}  -> xdg_activation_v1#17.activate("c74447bffb17e98a846b73a3a815d285", wl_surface#43)
```

and activation can succeed.

This manifests when the user didn't interact with the application after
startup as then the serials are all `0` and the application thus can't
request a valid token on their own.

Closes: #8022
Signed-off-by: Guido Günther <agx@sigxcpu.org>